+++ lib/responsive_spacing_v2.dart
@@ -1,3 +1,6 @@
 library responsive_spacing_v2;
 
 export 'src/responsive_spacing_v2.dart';
+export 'src/widget/responsive_spacer_v2.dart';
+export 'src/provider/responsive_spacing_provider_v2.dart';
+export 'src/core/responsive_spacing_v2.dart';